modified:   core/arms.hpp 	modified:   core/particle_smoother.hpp 	modified:   rng/prng.hpp 	modified:   rng/prng_engine.hpp corrected a non-return bug in inv_int_exp_y function in arms.hpp modified the numerical computation in inv_int_evp_y in arms.hpp

+++ rng/prng_engine.hpp
@@ -220,7 +220,7 @@ namespace sitmo {
     void discard(uint64_t z)
     {
       // check if we stay in the current block
-      if (z < 8 - _o_counter) {
+      if (z < (uint64_t)(8 - _o_counter)) {
 	_o_counter += static_cast<unsigned short>(z);
 	return;
       }
